Bound, harden, and back off the format cache after review
- Cap the cache at 2048 entries (evicting oldest-verified first) and
  write compact JSON: the whole document is rewritten and fsynced on the
  track-switching path, so it must stay bounded.
- Defer re-verification when a catalog lookup returns nothing so a
  delisted or unmatchable track no longer re-verifies on every playback.
- Run Clear() off the WPF UI thread; it serializes and fsyncs under the
  store lock and can block behind an in-flight background store.
- Reject store-backed resolvers in FormatCacheVerificationService: one
  would supersede its own compare-and-swap and silently drop updates.
- Keep valid cache entries when one persisted entry is malformed instead
  of rebuilding the whole cache.
- Treat future verification timestamps (clock skew) as due.
- Snap refresh-days in long arithmetic; int.MinValue inverted the snap.
- Unwedge the toast queue when Show() throws; coalesce queued cache
  toasts per track and cap them at the newest three.
- Reset the read-failure log latch on Clear(); sweep orphaned .tmp files.

The transient-unreadable tests now pin the actual regression (verified
to fail with the guard reverted), the cache generation assertions
compare specific non-zero values, and new coverage spans eviction,
per-entry skip, touch backoff, the fresh-entry no-verification negative,
toast cap/coalesce, and parallel store/clear/touch consistency.

+++ src/Models/AppSettings.cs
@@ -88,8 +88,10 @@ public sealed class AppSettings
             return days;
         }
 
+        // Distance in long: with an int, days near int.MinValue (a corrupt settings file) overflows
+        // Math.Abs and inverts the snap.
         return SupportedFormatCacheRefreshDays
-            .OrderBy(option => Math.Abs(option - days))
+            .OrderBy(option => Math.Abs((long)option - days))
             .ThenBy(option => option)
             .First();
     }
